Add button to reset checkbox defaults
Function `select_default_columns` now changes the columns in the table themselves, as well as the checkboxes.

Uses `.prop` instead of `.attr` because the latter has issues re-checking an unchecked box -- see https://stackoverflow.com/a/36185922.

I just copied the HTML/CSS for existing buttons - this should probably be changed.