Stop map pins from highlighting in white
Text for map pins on individual strain pages would highlight in white, which would blend into the background.  This is because the map was contained in a `list-group-item` class div, which has that behavior.

I've changed the class to `list-group-item-img`, to keep the class name as similar as possible and to avoid changing the CSS.