Clean up & comment fetch_orthologs
NOTE: In the original code, idx started from 1 (set to 0 and incremented at start of loop). In new code, it is zero-indexed. This doesn't affect anything, since it is only used to print the progress update & quit early for mock data, but it is still strictly speaking a change in operation.