Ack or nack Pub/Sub push message based on job completion
If job is not yet complete, send a 100 Continue response to nack the message, which will cause Pub/Sub to resubmit the message later.

I chose 100 because it's not an error, but it counts as a nack for Pub/Sub.  I want to reflect that the status update function ran properly, but that the client should continue to send requests.