Abstract Runner class away from specific Report
Instances of the `Runner` class or its subclasses are considered unique up to `data_id`.
For the tool jobs, this is the data hash; for database operations, this is the operation name (e.g. `TEST_ECHO`).

To run a job, a specific report must be provided.  This will fill out the template provided by the runner class with variables specific to that report.