Handle status check in Runner class
Runner now includes abstract method `check_status`, which can be used to check the status of a specific execution of the job.

The status check also handles updating the `PipelineOperation` record, so the task `/status` route doesn't have to explicitly update this anymore.
More concretely, the code for updating the record has been moved into the `GCPRunner` and `GCPCloudRunRunner` classes.

Doing things this way has highlighted the possibility of job name collisions. Since job names are composed of `{kind}-{data_id}`,
it's technically possible for two different jobs to have the same name if the `kind` of one is a prefix of the other.
This isn't possible with any of our currently existing `kind`s, and in fact is mitigated if the `data_id` cannot contain hyphens.
Some of the code in this commit is to try and reasonably future-proof this solution for just this problem.