Refactor TaskStatus to JobStatus
The status is more accurately a property of the pipeline as a whole, rather than the task specifically.

This has the nice effect of moving the definition out to the `services` level, so we no longer have to locally import the status in Entity methods to avoid circular imports.

Also added CSS color support for the new `CREATED` value, between report creation & task submission.