Get rid of tool-specific Runner subclasses
Parsing the `Report` to get `Runner.run` params now happens in the appropriate `JobPipeline` subclass. This lets the `Runner` class stay as agnostic to the specific job it's running as possible.

Still TODO: Parsing GCP operation names is now a bit trickier, since we can't rely on the class's `kind` to decide how to handle the special `DatabaseOperation` enum case.