Update cache checks for existing reports / submissions by other users
Now checks both for results and for another `COMPLETE` report, to make sure the results are valid.
This generalizes the existing behavior of `Nemascan` to all job types.

This might introduce a race condition if one job is submitted right as the job it's supposed to link to is completed...