Fix: Distinguish upload from string vs. file using LocalFile object
Moved `LocalFile` from `site-v2` module to the CaeNDR package.

Subclasses `os.PathLike`, which means it can be used directly in `open()`.
By passing along the full object instead of just the filename, we can check the type and decide how to upload that way -- as a file or as a string.