Comment out existing input check
I'm not sure what the best solution is for this - I think it won't hurt to re-upload, but it feels a bit odd to upload the file again if it already exists.

But, if there was an error uploading it the first time (or if it's somehow been changed/corrupted/etc), then all future jobs would fail.

We could download the file in GCP and check if the two are equal?