Move completion check -> status update to Indel Finder & Heritability pipeline classes
I'm not sure if this check holds true for NemaScan, since in principal we don't know when the report file is created relative to the other output files.
I'm pretty sure it should be the last one created, but I don't think it's good practice to implicitly rely on that.

As for the other tools, since they each only have one output file, we can be more confident that the job is done if the output exists.