Load tables w new `TableConfig` and `LocalDatastoreFile` classes
Table configurations are now stored in global objects of class `TableConfig`, which store:
- SQL table
- Parsing script
- Set of files required

Files are managed with `LocalDatastoreFile` class, which caches a GCP file locally and manages downloading/zipping.
These are in turn built by `LocalDatastoreFileTemplate` object, which store `TokenizedString`s and fill them out for a given species as needed.