Better error checking if files could not be retrieved
The existing `files and ...` guards work when `files is None`, but if `files` is an empty object `{}`,
then `{} and None` will be `{}`.  This would pass an empty object to `render_ext_markdown` as the `url`,
which in turn wasn't properly caught (since that function would only guard against `None`).

Changes:
- Better guard for files: `files or {}` will fallback to an empty dict, then `.get` will retrieve `None` since no fields exist.
- Better guard for empty "url"s: now checks against `bool` conversion, which catches `None` as well as `''` and `{}`.