Add explicit type hints to `lookup_report` pipeline objects
Lets the type checker / hinter access report-specific fields, which is most helpful right now for Indel Finder and Nemascan.

I attempted to look into e.g. `TypeVar` as a solution on `lookup_report`, but couldn't get it to produce the desired effect.  The type hint solution is manual, but it works exactly as desired.