Allow trait names to be stored along with file IDs in report
If one of the traits is from a bulk file, the file ID isn't enough to uniquely specify it. In this case, use the `trait_X_name` prop to store the unique name of the trait

TODO: The code for storing/retrieving names is a bit complicated because I'm not sure the best way to handle it for non-bulk files, i.e. cases where the file ID is enough to uniquely specify the trait. Starting to wonder if dataset + unique trait name is the better unique key?