Start adding code to filter by species
Idea would be that when one trait is selected, the selection modal for the other is filtered down to the same species.
However, I don't think this is straightforward yet -- there's no way right now to clear a selection, so once e.g. a C. elegans trait is chosen, the other menu will only ever show C. elegans traits.