Store trait names in report without checking `TraitFile` object
Since we're duplicating the trait name from the TraitFile, it would be nice to validate it, but if we do so every time we use the param,
we spam the datastore with requests and cause certain pages (namely my-results and all-results) to take a long time to load.

For now, I'm commenting out these functions.  Would like to return to as part of re-vamping the Entity class.

(This is in line with e.g. how we currently store `Container` fields in reports -- duplicate without checking directly.)