Fix announcement polling / flashing with context processor
Check for announcements in a Jinja template context processor, which means we will only flash messages when rendering a Jinja template. This automatically handles responses that don't return templates (i.e. ones that don't clear the flashed message queue), so messages shouldn't bleed over into the next request.

This also fixes an offset bug - we have to call `flash` before the template is rendered, otherwise announcements won't be queued in time and will be applied to the *next* page load.