Rename genome browser track `name` to `display_name` (SEE NOTE)
Having a custom `name` prop overwrites the GCP unique name, which is also saved as `.name` on `Entity` objects.
This change makes the GCP name available on the Entity.

We can phase this into the existing site as follows:
- Replace the existing `BrowserTrack` entities with new copies where:
  - The key is a "Custom Name" instead of a "Numeric ID"
  - The fields `name` and `display_name` both exist
- Push the changes from this commit / branch
- Delete the now-extraneous `name` fields

This is because, if we don't register `name` as a custom prop, that column in GCP will be ignored when constructing the `Entity`, and therefore it will not overwrite the GCP Custom Name